requirements.txt
googlesearch-python
pandas
numpy
datetime
bs4
newspaper3k
requests
nltk
gensim
xlsxwriter
matplotlib
spacy
textblob
wordcloud
pyLDAvis
openpyxl
networkx
textacy
people_also_ask
torch
librosa
transformers
pydub
seaborn
ffmpeg-python
html_table_parser
trafilatura